Resources aplication

+++ src/main/java/ifpb/edu/br/pj/ifpbichos/model/repository/ComissionMemberRepository.java
@@ -8,6 +8,6 @@ import org.springframework.stereotype.Repository;
 import ifpb.edu.br.pj.ifpbichos.model.entity.ComissionMember;
 
 @Repository
-public interface ComissionMemberRepository extends JpaRepository<ComissionMember, Integer>{
+public interface ComissionMemberRepository extends JpaRepository<ComissionMember, Long>{
 	public boolean existsByLogin(String Login);
 }
